requirements.txt
pandas
requests
sqlalchemy
python-dotenv
psycopg2-binary
google-genai
